requirements.txt
slack-bolt>=1.18.0
anthropic>=0.39.0
supabase>=2.9.0
python-dotenv>=1.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
requests>=2.31.0
